+++ RockWeb/Blocks/Administration/SystemInfo.ascx.cs
@@ -177,15 +177,20 @@ namespace RockWeb.Blocks.Administration
             // Delete all cached files
             try
             {
-                var dirInfo = new DirectoryInfo( Path.Combine( webAppPath, "App_Data/Cache" ) );
-                foreach ( var childDir in dirInfo.GetDirectories() )
+                var cachePath = Path.Combine( webAppPath, "App_Data/Cache" );
+                if ( Directory.Exists( cachePath ) )
                 {
-                    childDir.Delete( true );
+                    var dirInfo = new DirectoryInfo( cachePath );
+                    foreach ( var childDir in dirInfo.GetDirectories() )
+                    {
+                        childDir.Delete( true );
+                    }
+                    foreach ( var file in dirInfo.GetFiles().Where( f => f.Name != ".gitignore" ) )
+                    {
+                        file.Delete();
+                    }
                 }
-                foreach ( var file in dirInfo.GetFiles().Where( f => f.Name != ".gitignore" ) )
-                {
-                    file.Delete();
-                }
+
                 msgs.Add( "Cached files have been deleted" );
             }
             catch ( Exception ex )
